Fix profile migration; fixes #86

+++ hook.php
@@ -252,15 +252,17 @@ function plugin_datainjection_migration_24_250($migration) {
 function plugin_datainjection_upgrade23_240($migration) {
    global $DB;
 
-   if ($DB->fieldExists('glpi_plugin_datainjection_profiles', 'ID')) {
-      $migration->changeField('glpi_plugin_datainjection_profiles', 'ID', 'id', 'autoincrement');
-      $migration->executeMigration();
+   if ($DB->tableExists('glpi_plugin_datainjection_profiles')) {
+      if ($DB->fieldExists('glpi_plugin_datainjection_profiles', 'ID')) {
+         $migration->changeField('glpi_plugin_datainjection_profiles', 'ID', 'id', 'autoincrement');
+         $migration->executeMigration();
+      }
+
+       PluginDatainjectionProfile::migrateProfiles();
+
+      //Drop profile table : no use anymore !
+      $migration->dropTable('glpi_plugin_datainjection_profiles');
    }
-
-    PluginDatainjectionProfile::migrateProfiles();
-
-    //Drop profile table : no use anymore !
-    $migration->dropTable('glpi_plugin_datainjection_profiles');
 }
 
 function plugin_datainjection_update131_14() {
